requirements.txt
discord.py==1.7.3
markovify==0.9.4
jishaku==2.4.0
toml==0.10.2